# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@freerangenotify.com
    name: API Support
  description: High-performance notification service with multi-channel delivery support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FreeRangeNotify API
  version: 0.1.9-alpha
paths:
...
  /v1/admin/playground/sse:
    post:
      description: Generate a temporary SSE playground session with a 30-minute TTL
        for testing real-time notifications
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an SSE playground
      tags:
      - Playground
  /v1/admin/playground/sse/{id}/send:
    post:
      consumes:
      - application/json
      description: Publish a test notification to an SSE playground session
      parameters:
      - description: Playground ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional message content (title, body, category, data)
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Send an SSE test message
      tags:
      - Playground
  /v1/admin/playground/webhook:
    post:
      description: Generate a temporary webhook receiver URL with a 30-minute TTL
        for testing
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a webhook playground
      tags:
      - Playground
...
  /v1/playground/{id}:
    get:
      description: Retrieve all received webhook payloads for a playground session
      parameters:
      - description: Playground ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get playground payloads
      tags:
      - Playground
    post:
      consumes:
      - application/json
      description: Public endpoint that captures incoming webhook payloads for a playground
        session
      parameters:
      - description: Playground ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Receive a webhook payload
      tags:
      - Playground
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Application API key for tenant-facing endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and your API key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"